Move SearchResults to context folder
